gettssmerger文件名去除特殊字符,merge_to_mp4删除文件夹更正

+++ m3u8download.py
@@ -107,4 +107,5 @@ def merge_to_mp4(name, delete=False):
                 print(f'\r{file} Merged! Total:{len(files)}', end="     ")
             if delete:
                 os.remove(file)
-        os.rmdir(f'./{name}/')
+        if delete:
+            os.rmdir(f'./{name}/')
